refactor(filters): move filter type definitions to filters/types package
