+++ filters/pipelines.go
@@ -1,12 +1,17 @@
 package filters
 
-import "fmt"
+import (
+	"fmt"
+
+	"dougdomingos.com/image-filters/filters/grayscale"
+	"dougdomingos.com/image-filters/filters/types"
+)
 
 // AvailableFilters maps a string identifier to its corresponding filter
 // pipeline.
-var AvaliableFilters = map[string]FilterPipeline{
+var AvaliableFilters = map[string]types.FilterPipeline{
 	"binarization":     BinarizationPipeline,
-	"grayscale":        GrayscalePipeline,
+	"grayscale":        grayscale.GrayscalePipeline,
 	"horizontal-flip":  HorizontalFlipPipeline,
 	"sobel":            SobelPipeline,
 	"sobel-grayscaled": SobelGrayscaledPipeline,
@@ -17,10 +22,10 @@ var AvaliableFilters = map[string]FilterPipeline{
 // GetFilterPipeline retrieves a filter pipeline by its name from the
 // AvailableFilters map. It returns the pipeline if found, or an error if the
 // specified name is not defined.
-func GetFilterPipeline(filterName string) (FilterPipeline, error) {
+func GetFilterPipeline(filterName string) (types.FilterPipeline, error) {
 	pipeline, exists := AvaliableFilters[filterName]
 	if !exists {
-		return FilterPipeline{}, fmt.Errorf("[ERROR]: Specified filter does not exist")
+		return types.FilterPipeline{}, fmt.Errorf("[ERROR]: Specified filter does not exist")
 	}
 
 	return pipeline, nil
